feat: add new client

+++ internal/handlers/role_handlers.go
@@ -1,6 +1,8 @@
 package handlers
 
 import (
+	"fmt"
+
 	"github.com/gofiber/fiber/v2"
 	"github.com/scharph/orda/internal/ports"
 )
@@ -50,7 +52,14 @@ func (h *RoleHandlers) Update(c *fiber.Ctx) error {
 	if err := c.BodyParser(&req); err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid role data"})
 	}
+
 	req.Id = c.Params("id")
+
+	role, err := h.roleService.GetById(c.Context(), req.Id)
+	if role.Name == "admin" {
+		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Cannot update admin role"})
+	}
+
 	res, err := h.roleService.Update(c.Context(), req.Id, req)
 	if err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": err.Error()})
@@ -60,9 +69,13 @@ func (h *RoleHandlers) Update(c *fiber.Ctx) error {
 
 func (h *RoleHandlers) Delete(c *fiber.Ctx) error {
 	id := c.Params("id")
+	role, err := h.roleService.GetById(c.Context(), id)
+	if role.Name == "admin" {
+		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Cannot update admin role"})
+	}
 	res, err := h.roleService.Delete(c.Context(), id)
 	if err != nil {
-		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Failed to delete role"})
+		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": fmt.Sprintf("Failed to delete role: %s", err)})
 	}
 	return c.Status(fiber.StatusOK).JSON(fiber.Map{"success": res})
 }
